merge ?

+++ main-be/api/users.py
@@ -1,5 +1,5 @@
 from flask import Blueprint, request, jsonify, abort
-from models import db, User, dateStr, app,LeadPayload
+from models import db, User, Task, dateStr, app,LeadPayload, Message, get_query_page_users, generate_datetime_id
 from flask import Flask, request, jsonify, session
 from flask_login import LoginManager, login_user, logout_user, login_required, current_user
 import datetime
@@ -27,63 +27,10 @@ def get_users():
         }
     })
 
-# @user_bp.route("/<string:user_id/lead", methods=["GET"])
-# def get_user_leadId(user_id):
-#     user = db.session.get(User,user_id)
-
-#     if not user:
-#         print("Unknown user")
-#         abort(404, description="Unknown user")
-
-#     return jsonify()
-    
-
-from collections import namedtuple
-def get_query_page_users(lead_id, page, limit, search):
-    if lead_id != 0:
-        lead = db.session.get(LeadPayload, lead_id)
-        if not lead:
-            Pagination = namedtuple('Pagination', ['total', 'pages'])
-            empty_pagination = Pagination(total=0, pages=0)
-            return [], empty_pagination
-        
-        query = lead.users.filter(
-            and_(
-                User.role_id > 0,
-                User.role_id < 100,
-            )
-        )
-    else:
-        # Lấy toàn bộ user khi lead_id == 0
-        query = User.query.filter(
-            and_(
-                User.role_id > 0,
-                User.role_id < 100,
-            )
-        )
-
-    print('User_query', lead_id, query.count())
-
-    if search:
-        query = query.filter(
-            (User.username.ilike(f"%{search}%")) | 
-            (User.fullName.ilike(f"%{search}%"))
-        )
-
-    split_length = func.array_length(func.regexp_split_to_array(User.fullName, ' '), 1)
-    last_name = func.split_part(User.fullName, ' ', split_length)
-
-    query = query.order_by(desc(User.updatedAt)).order_by(last_name, User.id)
-
-    pagination = query.paginate(page=page, per_page=limit, error_out=False)
-    users = [c.to_dict() for c in pagination.items]
-
-    return users, pagination
-
 @user_bp.route("/", methods=["POST"])
 def create_user():
     data = request.get_json()
-    print('USER', data)
+    # print('USER', data)
     new_user = User.create_item(data)
         
     db.session.add(new_user)
@@ -95,20 +42,105 @@ def create_user():
         raise
 
     db.session.refresh(new_user)
+
+    task = Task(
+            id=generate_datetime_id(),
+            type="salary",
+            assign_ids=[new_user.id]
+        )
+    db.session.add(task)
     
-    return jsonify(new_user.to_dict()), 201
+    return jsonify(new_user.tdict()), 201
+
+
+
+
+@user_bp.route("/<string:user_id>/password", methods=["PUT"])
+def update_lead_user_password(user_id):
+    data = request.get_json()
+    print(data)
+
+    user = db.session.get(User, user_id)
+
+    if not user:
+        print("Lead not found")
+        abort(404, description="Lead not found")
+
+    if data.get("old_password") != user.password:
+        print("Mật khẩu cũ không đúng")
+        abort(404, description="Mật khẩu cũ không đúng")
+    
+    # id = db.Column(db.Integer, primary_key=True)
+    # name = db.Column(db.String(120), nullable=False)
+    if data.get("fullName"):
+        user.fullName = data.get("fullName")
+
+    if data.get("username"):
+        user.username = data.get("username")
+
+    if data.get("password"):
+        user.password = data.get("password")
+        
+
+    # if data.get("company"):
+    #     user.company = data.get("company")
+
+    # if data.get("address"):
+    #     user.address = data.get("address")
+
+    # if data.get("email"):
+    #     user.email = data.get("email")
+
+    # if data.get("phone"):
+    #     user.phone = data.get("phone")
+
+    # if data.get("description"):
+    #     user.description = data.get("description")
+
+    # if data.get("industry"):
+    #     user.industry = data.get("industry")
+
+    db.session.commit()
+
+    # companySize = db.Column(db.String(50), nullable=False)
+    # expiredAt = db.Column(db.DateTime, default=datetime.datetime.utcnow)
+    # balance_amount = db.Column(db.Float)
+
+    # balance_amount = db.Column(db.Float)
+
+    db.session.refresh(user)
+    return jsonify(user.tdict()), 201
+
+@user_bp.route("/<string:user_id>/check-password", methods=["POST"])
+def check_password_lead(user_id):
+    data = request.get_json()
+
+    user = db.session.get(User, user_id)
+
+    if not user:
+        print("Lead-user not found")
+        abort(404, description="Lead-user not found")
+
+    if data.get("old_password") != user.password:
+        print("Mật khẩu cũ không đúng")
+        abort(404, description="Mật khẩu cũ không đúng")
+
+    return jsonify({'message':'right password'}), 200
+
+
+
 
 @user_bp.route("/<string:id>", methods=["GET"])
 def get_user_detail(id):
     user = db.session.get(User, id)
     if not user:
         abort(404, description="user not found")
-    return jsonify(user.to_dict())
+    return jsonify(user.tdict())
 
 @user_bp.route("/<string:id>", methods=["PUT"])
 def update_user(id):
     data = request.get_json()
-    print('PUT user', data)
+    # print('PUT user', data)
     user = db.session.get(User, id)
     if not user:
         return jsonify({"error": "user not found"}), 404
@@ -140,8 +172,7 @@ def update_user(id):
         db.session.rollback()
         return jsonify({"error": str(e)}), 500
     
-    return jsonify(user.to_dict()), 200
-
+    return jsonify(user.tdict()), 200
 
 
 
@@ -151,8 +182,18 @@ def update_user(id):
 @user_bp.route("/<string:id>", methods=["DELETE"])
 def delete_user(id):
     user = db.session.get(User, id)
-    if user:
-        db.session.delete(user)
-    db.session.commit()
+    
+    if not user:
+        print("No user", id)
+        abort(404, description = "No user")
+    else:
+        Message.query.filter_by(user_id=user.id).delete(synchronize_session=False)
+        db.session.commit()  # Đảm bảo các message được xóa trước, tránh vi phạm khoá ngoại
+
+        # Lấy user và xóa
+        user = db.session.get(User, user.id)
+        if user:
+            db.session.delete(user)
+            db.session.commit()
 
     return jsonify({"message": "User deleted"}), 200
